backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
openai==1.45.0
tavily-python==0.5.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
